\name{cgdsr-processURL}
\alias{cgdsr-processURL}
\alias{processURL}
\alias{processURL.CGDS}
\title{Internal methods for CGDS library.}
\description{These methods should not be invoked by the user.}
\author{
<jacobsen@cbio.mskcc.org>
}
\references{
  CGDS public server at MSKCC: http://cbio.mskcc.org/cgds-public/
}
\seealso{
\code{\link{cgdsr}},\code{\link{CGDS}}
}
