\name{cgdsr-processURL}
\alias{cgdsr-processURL}
\alias{processURL}
\alias{processURL.CGDS}
\title{Internal methods for CGDS library.}
\description{These methods should not be invoked by the user.}
\author{
<jacobsen@cbio.mskcc.org>
}
\references{
 cBio Cancer Genomics Portal: \url{http://www.cbioportal.org/}
}
\seealso{
\code{\link{cgdsr}},\code{\link{CGDS}}
}
